+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Add `editor.line_number_scale` setting for smaller line numbers
+
 ## Fixed
 
 - Prevent table cutoff in markdown preview (#332) by @kmkkiii